+++ bin/services/workspace.js
@@ -3,7 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.getWorkingTickets = exports.addWorkspace = exports.listWorkspaces = exports.checkWorkspace = undefined;
+exports.updateWorkspace = exports.getWorkingTickets = exports.addWorkspace = exports.listWorkspaces = exports.checkWorkspace = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
@@ -129,6 +129,37 @@ let getWorkingTickets = (() => {
   };
 })();
 
+let updateWorkspace = (() => {
+  var _ref6 = _asyncToGenerator(function* (workspace) {
+    yield checkWorkspace();
+
+    let workspaces;
+
+    const projectPath = process.cwd();
+    const workspacesPath = _path2.default.join(projectPath, WORKSPACE_FILENAME);
+
+    try {
+      yield (0, _fs.access)(workspacesPath);
+
+      const fileContent = yield (0, _fs.readFile)(workspacesPath, "utf8");
+
+      workspaces = JSON.parse(fileContent);
+    } catch (err) {
+      workspaces = {};
+    }
+
+    workspaces = _extends({}, workspaces, workspace);
+
+    yield saveWorkspace(workspaces);
+
+    return workspace;
+  });
+
+  return function updateWorkspace(_x4) {
+    return _ref6.apply(this, arguments);
+  };
+})();
+
 var _path = require("path");
 
 var _path2 = _interopRequireDefault(_path);
@@ -149,3 +180,4 @@ exports.checkWorkspace = checkWorkspace;
 exports.listWorkspaces = listWorkspaces;
 exports.addWorkspace = addWorkspace;
 exports.getWorkingTickets = getWorkingTickets;
+exports.updateWorkspace = updateWorkspace;